Chain 함수 완성. difficulty 증가하지 않는 버그 있음.

+++ jaeseung-blockchain/src/constants.rs
@@ -1,8 +1,8 @@
 // 블록 1 묶음의 블록 개수
-pub const DIFFICULTY_ADJUSTMENT_INTERVAL_COUNT: u16 = 10;
+pub const DIFFICULTY_ADJUSTMENT_INTERVAL_COUNT: u32 = 10;
 
 // 블록 1 묶음이 생성 되기까지의 예상 시간(분)
-pub const BLOCK_GENERATION_INTERVAL_MINUTE: u16 = 10;
+pub const BLOCK_GENERATION_INTERVAL_MINUTE: u32 = 10;
 
 // 블록 1 개가 생성 되기까지의 예상 시간(밀리초)
-pub const BLOCK_GENERATION_MILLIS: u16 = 60000;
+pub const BLOCK_GENERATION_MILLIS: u32 = 60000;
